requirements.txt
pynetdicom==2.0.2
pydicom==2.4.4
PyYAML==6.0.1
psycopg2-binary==2.9.9
